+++ tests/lib/js/static_interop_test/js_default_test.dart
@@ -12,7 +12,7 @@ import 'dart:js_interop';
 
 import 'package:expect/minitest.dart';
 
-import 'js_default_other_library.dart' as other;
+import 'js_default_with_namespaces.dart' as namespace;
 
 @JS()
 external void eval(String code);
@@ -29,6 +29,20 @@ class SimpleObject {
 }
 
 extension SimpleObjectExtension on SimpleObject {
+  external JSNumber get initialArguments;
+  external JSNumber twoOptional([JSNumber n1, JSNumber n2]);
+  external JSNumber oneOptional(JSNumber n1, [JSNumber n2]);
+}
+
+@JS('SimpleObject')
+extension type SimpleObject2._(JSObject _) implements JSObject {
+  external factory SimpleObject2();
+  external factory SimpleObject2.twoOptional([JSNumber n1, JSNumber n2]);
+  external factory SimpleObject2.oneOptional(JSNumber n1, [JSNumber n2]);
+
+  external static JSNumber twoOptionalStatic([JSNumber n1, JSNumber n2]);
+  external static JSNumber oneOptionalStatic(JSNumber n1, [JSNumber n2]);
+
   external JSNumber get initialArguments;
   external JSNumber twoOptional([JSNumber n1, JSNumber n2]);
   external JSNumber oneOptional(JSNumber n1, [JSNumber n2]);
@@ -75,30 +89,77 @@ void testCurrentLibrary() {
 
   expect(1, s.oneOptional(4.0.toJS).toDartInt);
   expect(2, s.oneOptional(4.0.toJS, 5.0.toJS).toDartInt);
+
+  // Test extension type factories.
+  expect(0, SimpleObject2.twoOptional().initialArguments.toDartInt);
+  expect(1, SimpleObject2.twoOptional(4.0.toJS).initialArguments.toDartInt);
+  expect(2,
+      SimpleObject2.twoOptional(4.0.toJS, 5.0.toJS).initialArguments.toDartInt);
+
+  expect(1, SimpleObject2.oneOptional(4.0.toJS).initialArguments.toDartInt);
+  expect(2,
+      SimpleObject2.oneOptional(4.0.toJS, 5.0.toJS).initialArguments.toDartInt);
+
+  // Test extension type static methods.
+  expect(0, SimpleObject2.twoOptionalStatic().toDartInt);
+  expect(1, SimpleObject2.twoOptionalStatic(4.0.toJS).toDartInt);
+  expect(2, SimpleObject2.twoOptionalStatic(4.0.toJS, 5.0.toJS).toDartInt);
+
+  expect(1, SimpleObject2.oneOptionalStatic(4.0.toJS).toDartInt);
+  expect(2, SimpleObject2.oneOptionalStatic(4.0.toJS, 5.0.toJS).toDartInt);
+
+  // Test extension type methods.
+  final s2 = SimpleObject2();
+  expect(0, s2.twoOptional().toDartInt);
+  expect(1, s2.twoOptional(4.0.toJS).toDartInt);
+  expect(2, s2.twoOptional(4.0.toJS, 5.0.toJS).toDartInt);
+
+  expect(1, s2.oneOptional(4.0.toJS).toDartInt);
+  expect(2, s2.oneOptional(4.0.toJS, 5.0.toJS).toDartInt);
 }
 
 void testOtherLibrary() {
   // Test top level methods.
-  expect(1, other.oneOptional(4.0.toJS).toDartInt);
-  expect(2, other.oneOptional(4.0.toJS, 5.0.toJS).toDartInt);
+  expect(1, namespace.oneOptional(4.0.toJS).toDartInt);
+  expect(2, namespace.oneOptional(4.0.toJS, 5.0.toJS).toDartInt);
 
   // Test factories.
-  expect(
-      1, other.SimpleObject.oneOptional(4.0.toJS).initialArguments.toDartInt);
+  expect(1,
+      namespace.SimpleObject.oneOptional(4.0.toJS).initialArguments.toDartInt);
   expect(
       2,
-      other.SimpleObject.oneOptional(4.0.toJS, 5.0.toJS)
+      namespace.SimpleObject.oneOptional(4.0.toJS, 5.0.toJS)
           .initialArguments
           .toDartInt);
 
   // Test static methods.
-  expect(1, other.SimpleObject.oneOptionalStatic(4.0.toJS).toDartInt);
-  expect(2, other.SimpleObject.oneOptionalStatic(4.0.toJS, 5.0.toJS).toDartInt);
+  expect(1, namespace.SimpleObject.oneOptionalStatic(4.0.toJS).toDartInt);
+  expect(2,
+      namespace.SimpleObject.oneOptionalStatic(4.0.toJS, 5.0.toJS).toDartInt);
 
   // Test extension methods.
-  final s = other.SimpleObject();
+  final s = namespace.SimpleObject();
   expect(1, s.oneOptional(4.0.toJS).toDartInt);
   expect(2, s.oneOptional(4.0.toJS, 5.0.toJS).toDartInt);
+
+  // Test extension type factories.
+  expect(1,
+      namespace.SimpleObject.oneOptional(4.0.toJS).initialArguments.toDartInt);
+  expect(
+      2,
+      namespace.SimpleObject.oneOptional(4.0.toJS, 5.0.toJS)
+          .initialArguments
+          .toDartInt);
+
+  // Test extension type static methods.
+  expect(1, namespace.SimpleObject.oneOptionalStatic(4.0.toJS).toDartInt);
+  expect(2,
+      namespace.SimpleObject.oneOptionalStatic(4.0.toJS, 5.0.toJS).toDartInt);
+
+  // Test extension type methods.
+  final s2 = namespace.SimpleObject();
+  expect(1, s2.oneOptional(4.0.toJS).toDartInt);
+  expect(2, s2.oneOptional(4.0.toJS, 5.0.toJS).toDartInt);
 }
 
 void main() {
@@ -127,5 +188,19 @@ void main() {
   }
   ''');
   testCurrentLibrary();
+  // Move the declarations to a namespace and delete the declarations on
+  // globalThis to make sure we incorporate the library prefix in
+  // invocation-level lowering.
+  eval('''
+  var library1 = {};
+  globalThis.library1 = library1;
+
+  library1.twoOptional = globalThis.twoOptional;
+  library1.oneOptional = globalThis.oneOptional;
+  delete globalThis.twoOptional;
+  delete globalThis.oneOptional;
+  library1.SimpleObject = globalThis.SimpleObject;
+  delete globalThis.SimpleObject;
+  ''');
   testOtherLibrary();
 }