+++ CHANGELOG.md
@@ -1,3 +1,13 @@
+## 3.3.4
+
+This is a patch release that:
+
+- Fixes an issue with JS interop in dart2wasm where JS interop methods that used
+  the enclosing library's `@JS` annotation were actually using the invocation's
+  enclosing library's `@JS` annotation. (issue [#55359])
+
+[#55359]: https://github.com/dart-lang/sdk/issues/55359
+
 ## 3.3.3 - 2024-03-27
 
 This is a patch release that: